Revision of Authorization
I remove DataAccessLayer and database to simplify the implementation and  add a checkClaim function in class Controller.

At initial , I assign claims to usesr in main program which can later be done by an user managament . Then I have the users try to access the feature which they are and are not allowed.