AuthorizationManagerTests
Updated HasEqualOrHigherPrivilege to HasHigherPrivilege

Created tests for FindHeight and HasHigherPrivilege in AuthorizationManager. 4 tests fail but one is a duplicate of FindHeight. Still trying to get them to pass,
Added to AuthorizationDemo to see results of FindHeight and they come out correct. But do not work in Tests for some reason. Might be something with entitiy framework